feat: 상품 등록

+++ backend/src/main/java/com/dging/dgingmarket/service/security/CustomUserDetailsService.java
@@ -16,6 +16,6 @@ public class CustomUserDetailsService implements UserDetailsService {
 
     @Override
     public UserDetails loadUserByUsername(String userPk) throws UsernameNotFoundException {
-        return userRepository.findById(Integer.parseInt(userPk)).orElseThrow(CEntityNotFoundException.CUserNotFoundException::new);
+        return userRepository.findById(Long.parseLong(userPk)).orElseThrow(CEntityNotFoundException.CUserNotFoundException::new);
     }
 }
